+++ CHANGELOG.md
@@ -10,9 +10,14 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ## [Unreleased]
 
-## 2026-07-22
+## 2026-07-24
 
 - yellowstone-grpc-geyser 14.2.2
+
+### Fixed
+
+- Fixed `yellowstone_grpc_geyser_concurrent_subscribe_per_subscriber_id` when `subscription_limit_enforce` is set to false.
+- The `BlockMeta` message is now handle and only sent to geyser loop before being delegated to the block-reconstruction loop, this ensure we see all account updates before block-meta since on block `freeze` new account update are emitted really late, so a race could happen and cause lost account updates.
 
 ### Misc